# tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "actjs",
    "description": "Self-hosted actor framework HTTP API",
    "version": "0.3.0"
  },
...
  "paths": {
...
    "/v1/classes/{name}/versions/{version}": {
      "patch": {
        "summary": "Deprecate a class version (admin)",
        "tags": [
          "classes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "def": {
                  "type": "object",
                  "shape": {
                    "deprecated": {
                      "def": {
                        "type": "literal",
                        "values": [
                          true
                        ]
                      },
                      "type": "literal",
                      "values": {}
                    },
                    "graceUntilMs": {
                      "def": {
                        "type": "optional",
                        "innerType": {
                          "def": {
                            "type": "number",
                            "checks": [
                              {
                                "def": {
                                  "type": "number",
                                  "check": "number_format",
                                  "abort": false,
                                  "format": "safeint"
                                },
                                "type": "number",
                                "minValue": -9007199254740991,
                                "maxValue": 9007199254740991,
                                "isInt": true,
                                "isFinite": true,
                                "format": "safeint"
                              },
                              {}
                            ]
                          },
                          "type": "number",
                          "minValue": 0,
                          "maxValue": 9007199254740991,
                          "isInt": true,
                          "isFinite": true,
                          "format": "safeint"
                        }
                      },
                      "type": "optional"
                    }
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "toJSONSchema",
            "required": true
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "name": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                },
                "version": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": null,
                  "minLength": 1,
                  "maxLength": null
                }
              }
            },
            "in": "path",
            "name": "def",
            "required": true
          },
          {
            "schema": "object",
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "in": "path",
            "name": "parse",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParse",
            "required": true
          },
          {
            "in": "path",
            "name": "parseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "spa",
            "required": true
          },
          {
            "in": "path",
            "name": "encode",
            "required": true
          },
          {
            "in": "path",
            "name": "decode",
            "required": true
          },
          {
            "in": "path",
            "name": "encodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "decodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecodeAsync",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "name": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "version": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "deprecated": {
                        "def": {
                          "type": "literal",
                          "values": [
                            true
                          ]
                        },
                        "type": "literal",
                        "values": {}
                      },
                      "graceUntilMs": {
                        "def": {
                          "type": "optional",
                          "innerType": {
                            "def": {
                              "type": "number",
                              "checks": []
                            },
                            "type": "number",
                            "minValue": null,
                            "maxValue": null,
                            "isInt": false,
                            "isFinite": true,
                            "format": null
                          }
                        },
                        "type": "optional"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}